Update citynames.py
For some reason, I had changed random.randint to random.uniform before I checked it in.  I had tested it with randint and it works.  There is a rounding error in the code if one uses uniform instead.  It happens when forcing it to generate city names beginning with X (and removing the special secondLetter dictionary) - X only had three letters after it, L, a, e.  e was never chosen.  With the secondLetter dictionary in place, it worked fine, but without it, XL appeared about 50% of the time.

Summary:  secondLetter caused better names.  randint uses the full extent of the frequencies.  Fixed randint with this checkin.